add doc on sigma. fix issue on double parsing

+++ K2Bridge/Factories/AggregateFactory.cs
@@ -319,7 +319,7 @@ internal static class AggregateFactory
             return extendedStatsAggregate;
         }
 
-        var sigma = Convert.ToInt32(columnMetadata[2]);
+        var sigma = Convert.ToDouble(columnMetadata[2]);
 
         var jObject = (JObject)row[columnName];
